Added node recreate(), synced submodule

+++ commXel_app/src/ap/xel_net/xel_net_node.hpp
@@ -43,8 +43,8 @@ private:
 
       sensor_msgs::BatteryState battery_state_topic;
       battery_state_topic.header.frame_id = (char*) "CommXel BatteryState";
-      battery_state_topic.header.stamp.sec = (int32_t)(nano_time_/1000000000);
-      battery_state_topic.header.stamp.nanosec = (uint32_t)(nano_time_%1000000000);
+      battery_state_topic.header.stamp.sec = (int32_t)(nano_time_/(uint64_t)1000000000UL);
+      battery_state_topic.header.stamp.nanosec = (uint32_t)(nano_time_%(uint64_t)1000000000UL);
 
       battery_state_topic.voltage = 1;
       battery_state_topic.current = 2;
